--- packages/twenty-zapier/src/test/utils/labelize.test.ts
@@ -1,7 +1,0 @@
-import { labelling } from '../../utils/labelling';
-
-describe('labelling', () => {
-  test('should label properly', () => {
-    expect(labelling('createdAt')).toEqual('Created At');
-  });
-});
